# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is the API documentation for the Access API.",
      "version": "1.0.0",
      "title": "Access API"
    },
    "host": "localhost:5002",
    "basePath": "/api",
...
    "paths": {
...
      "/post": {
        "post": {
          "tags": ["post"],
          "summary": "Create a new post",
          "description": "Creates a new post",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "topicId": { "type": "integer" },
                  "userId": { "type": "integer" },
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "image": { "type": "string" }
                },
                "required": ["topicId", "userId", "title", "content", "image"]
              }
            }
          ],
          "responses": {
            "200": { "description": "Post created successfully" },
            "400": { "description": "Invalid input" }
          }
        },
        "get": {
          "tags": ["post"],
          "summary": "Get all posts",
          "description": "Retrieves a list of all posts",
          "responses": {
            "200": { "description": "List of posts" },
            "404": { "description": "No posts found" }
          }
        }
      },
      "/post/{post_id}": {
        "get": {
          "tags": ["post"],
          "summary": "Get post by ID",
          "description": "Retrieves details of a specific post",
          "parameters": [
            {
              "name": "post_id",
              "in": "path",
              "required": true,
              "type": "integer",
              "format": "int64",
              "description": "ID of the post to return"
            }
          ],
          "responses": {
            "200": { "description": "Post details" },
            "404": { "description": "Post not found" }
          }
        }
      },
      "/post/like": {
        "put": {
          "tags": ["post"],
          "summary": "Like a post",
          "description": "Likes a specific post",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Post ID to like",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "postId": { "type": "integer" }
                },
                "required": ["postId"]
              }
            }
          ],
          "responses": {
            "200": { "description": "Post liked successfully" },
            "400": { "description": "Invalid input" }
          }
        }
      },
...
      "/comment/post": {
        "post": {
          "tags": ["comment"],
          "summary": "Add a comment to a post",
          "description": "Adds a comment to a post",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Comment object that needs to be added",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "postId": { "type": "integer" },
                  "userId": { "type": "integer" },
                  "content": { "type": "string" }
                },
                "required": ["postId", "userId", "content"]
              }
            }
          ],
          "responses": {
            "200": { "description": "Comment added successfully" },
            "400": { "description": "Invalid input" }
          }
        }
      },
...
      "/search/post/{query}": {
        "get": {
          "tags": ["search"],
          "summary": "Search posts",
          "description": "Searches for posts matching the query",
          "parameters": [
            {
              "name": "query",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "Query string to search posts"
            }
          ],
          "responses": {
            "200": { "description": "Search results" },
            "404": { "description": "No posts found" }
          }
        }
      },
...
    }
  }